+++ packages/editor/web/src/RichContentEditor/RichContentEditor.tsx
@@ -147,6 +147,7 @@ export interface RichContentEditorProps extends PartialDraftEditorProps {
     updateEditorStateCallback: (editorState: EditorState) => void
   ) => DraftEditorProps['handleReturn'];
   tablePluginMenu?: boolean;
+  callOnChangeOnNewEditorState?: boolean;
   /** This is a legacy API, chagnes should be made also in the new Ricos Editor API **/
 }
 
@@ -497,7 +498,11 @@ class RichContentEditor extends Component<RichContentEditorProps, State> {
     }
     // TODO: new editor state should become initialContentState?
     if (nextProps.editorState && this.props.editorState !== nextProps.editorState) {
-      this.setState({ editorState: nextProps.editorState });
+      if (this.props.callOnChangeOnNewEditorState) {
+        this.updateEditorState(nextProps.editorState);
+      } else {
+        this.setState({ editorState: nextProps.editorState });
+      }
     }
     if (this.props.theme !== nextProps.theme) {
       this.setState({ theme: nextProps.theme });